test: [M3-8313] - Move OBJ Multicluster Cypress tests to `ObjectStorageMulticluster` (#11484)
* Move OBJ multicluster tests from `objectStorage` to `objectStorageMulticluster` dir

* Added changeset: Improve organization of Object Storage and Object Storage Multicluster tests
