+++ packages/manager/.changeset/pr-11484-tests-1736279319564.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tests
+---
+
+Improve organization of Object Storage and Object Storage Multicluster tests ([#11484](https://github.com/linode/manager/pull/11484))
