+++ packages/manager/cypress/e2e/core/objectStorageMulticluster/bucket-details-multicluster.spec.ts
@@ -7,8 +7,10 @@ import {
   regionFactory,
 } from 'src/factories';
 import { randomLabel } from 'support/util/random';
+import { mockGetBucket } from 'support/intercepts/object-storage';
+import { mockGetRegions } from 'support/intercepts/regions';
 
-describe('Object Storage Gen 1 Bucket Details Tabs', () => {
+describe('Object Storage Multicluster Bucket Details Tabs', () => {
   beforeEach(() => {
     mockAppendFeatureFlags({
       objMultiCluster: true,
@@ -31,11 +33,17 @@ describe('Object Storage Gen 1 Bucket Details Tabs', () => {
   });
 
   describe('Properties tab without required capabilities', () => {
+    /*
+     * - Confirms that Gen 2-specific "Properties" tab is absent when OBJ Multicluster is enabled.
+     */
     it(`confirms the Properties tab does not exist for users without 'Object Storage Endpoint Types' capability`, () => {
-      const { region, label } = mockBucket;
+      const { label } = mockBucket;
+
+      mockGetBucket(label, mockRegion.id);
+      mockGetRegions([mockRegion]);
 
       cy.visitWithLogin(
-        `/object-storage/buckets/${region}/${label}/properties`
+        `/object-storage/buckets/${mockRegion.id}/${label}/properties`
       );
 
       cy.wait(['@getFeatureFlags', '@getAccount']);
@@ -47,8 +55,6 @@ describe('Object Storage Gen 1 Bucket Details Tabs', () => {
 
       // Confirm that "Properties" tab is absent.
       cy.findByText('Properties').should('not.exist');
-
-      // TODO Confirm "Not Found" notice is present.
     });
   });
 });
